requirements.txt
streamlit
Pillow
transformers
joblib
torch